CM-857: Rearranged code to allow REST cache of search results
